// src/api/generated/api.ts
export interface UserDto {
    'id'?: string;
    'email'?: string;
    'firstName'?: string;
    'lastName'?: string;
    'roles'?: Array<string>;
    'avatarUrl'?: string;
    'createdAt'?: string;
    'updatedAt'?: string;
}